feature #445 [CoreBundle]: Move all expression language related classes to folder expression
