+++ src/Bundle/CoreBundle/Expression/ValidationExpressionChecker.php
@@ -6,7 +6,7 @@
  * Time: 16:18
  */
 
-namespace UniteCMS\CoreBundle\Security;
+namespace UniteCMS\CoreBundle\Expression;
 
 use Symfony\Component\ExpressionLanguage\SyntaxError;
 use UniteCMS\CoreBundle\Entity\FieldableContent;
